Further broke everything down into components leaving only the App component left.
